# src/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API MemoBelc",
    "description": "Documentação da API MemoBelc. Use **Authorize** para informar o token Bearer e testar rotas protegidas."
  },
  "host": "127.0.0.1:5000",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token JWT: use o valor 'Bearer <seu_token>'"
    }
  },
  "security": [],
...
  "paths": {
...
    "/books/list": {
      "get": {
        "tags": ["Books"],
        "summary": "Livros disponíveis para o usuário",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Lista de livros" } }
      }
    },
    "/books/get/{book_id}": {
      "get": {
        "tags": ["Books"],
        "summary": "Buscar livro por ID",
        "security": [{ "Bearer": [] }],
        "parameters": [{ "name": "book_id", "in": "path", "required": true, "type": "string" }],
        "responses": { "200": { "description": "Livro" }, "403": { "description": "Sem acesso" }, "404": { "description": "Não encontrado" } }
      }
    },
    "/books/purchase": {
      "post": {
        "tags": ["Books"],
        "summary": "Comprar/adicionar livro à biblioteca",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["book_id"],
              "properties": { "book_id": { "type": "string" } }
            }
          }
        ],
        "responses": { "200": { "description": "Livro adicionado" }, "404": { "description": "Livro não encontrado" } }
      }
    },
    "/books/admin/create": {
      "post": {
        "tags": ["Books"],
        "summary": "Criar livro (admin)",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["titulo", "idioma", "nivel"],
              "properties": {
                "titulo": { "type": "string" },
                "idioma": { "type": "string" },
                "nivel": { "type": "string" }
              }
            }
          }
        ],
        "responses": { "201": { "description": "Livro criado" }, "403": { "description": "Apenas admin" } }
      }
    },
    "/books/admin/update/{book_id}": {
      "put": {
        "tags": ["Books"],
        "summary": "Atualizar livro (admin)",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "book_id", "in": "path", "required": true, "type": "string" },
          { "name": "body", "in": "body", "schema": { "type": "object" } }
        ],
        "responses": { "200": { "description": "Atualizado" }, "403": { "description": "Apenas admin" }, "404": { "description": "Não encontrado" } }
      }
    },
    "/books/admin/delete/{book_id}": {
      "delete": {
        "tags": ["Books"],
        "summary": "Excluir livro (admin)",
        "security": [{ "Bearer": [] }],
        "parameters": [{ "name": "book_id", "in": "path", "required": true, "type": "string" }],
        "responses": { "200": { "description": "Excluído" }, "403": { "description": "Apenas admin" }, "404": { "description": "Não encontrado" } }
      }
    },
    "/books/admin/list": {
      "get": {
        "tags": ["Books"],
        "summary": "Listar todos os livros (admin)",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Lista de livros" }, "403": { "description": "Apenas admin" } }
      }
    },
    "/books/admin/book/{book_id}": {
      "get": {
        "tags": ["Books"],
        "summary": "Livro com usuários (admin)",
        "security": [{ "Bearer": [] }],
        "parameters": [{ "name": "book_id", "in": "path", "required": true, "type": "string" }],
        "responses": { "200": { "description": "Livro e usuários" }, "403": { "description": "Apenas admin" }, "404": { "description": "Não encontrado" } }
      }
    },
    "/books/admin/assign": {
      "post": {
        "tags": ["Books"],
        "summary": "Atribuir livro a usuário (admin)",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["user_id", "book_id"],
              "properties": { "user_id": { "type": "string" }, "book_id": { "type": "string" } }
            }
          }
        ],
        "responses": { "200": { "description": "Atribuído" }, "403": { "description": "Apenas admin" }, "400": { "description": "user_id/book_id obrigatórios" } }
      }
    },
...
  }
}